requirements.txt
boto3>=1.34.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
